# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Welcome",
      "description": "歡迎訊息 router"
    },
    {
      "name": "Users",
      "description": "使用者 router"
    },
    {
      "name": "Todos",
      "description": "代辦事項 router"
    },
    {
      "name": "Upload",
      "description": "上傳 router"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Welcome"
        ],
        "description": "歡迎訊息",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "使用者註冊",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "使用者註冊",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/users/sign_in": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "使用者登入",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UsersSingIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/sign_out": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "使用者登出",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "description": "TODO 列表",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "description": "TODO 新增",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "body",
            "description": "JWT Token",
            "schema": {
              "$ref": "#/definitions/AddUserData"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "tags": [
          "Todos"
        ],
        "description": "TODO 修改",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          },
          {
            "name": "todo",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "description": "TODO 刪除",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/upload/": {
      "get": {
        "tags": [
          "Upload"
        ],
        "description": "上傳檔案確認頁",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Upload"
        ],
        "description": "上傳檔案",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "JWT Token",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Users": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "example": "string"
            },
            "nickname": {
              "type": "string",
              "example": "string"
            },
            "password": {
              "type": "string",
              "example": "string"
            }
          }
        }
      }
    },
    "UsersSingIn": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "example": "string"
            },
            "password": {
              "type": "string",
              "example": "string"
            }
          }
        }
      }
    },
    "AddUserData": {
      "type": "object",
      "properties": {
        "todo": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "example": "string"
            },
            "fileUrl": {
              "type": "string",
              "example": "string"
            }
          }
        }
      }
    },
    "UpdateUserData": {
      "type": "object",
      "properties": {
        "todo": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "example": "string"
            },
            "fileUrl": {
              "type": "string",
              "example": "string"
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {}
  }
}